ark_ec::hashing::curve_maps
Module wb

Structs

IsogenyMap

IsogenyMap defines an isogeny between curves of form Phi(x, y) := (a(x), b(x)*y). The xcoordinate of the codomain point only depends on thex-coordinate of the domain point, and the y-coordinate of the codomain point is a multiple of the y-coordinate of the domain point. The multiplier depends on the x`-coordinate of the domain point. All isogeny maps of curves of short Weierstrass form can be written in this way. See [[Ga18]]. Theorem 9.7.5 for details.

WBMap

Traits

WBConfig

Trait defining the necessary parameters for the WB hash-to-curve method for the curves of Weierstrass form of: of y^2 = x^3 + a*x + b where b != 0 but a can be zero like BLS-381 curve. From [[WB2019]]