ark_ec::hashing::map_to_curve_hasher
Trait MapToCurve

pub trait MapToCurve<T: CurveGroup>: Sized {
    // Required methods
    fn new() -> Result<Self, HashToCurveError>;
    fn map_to_curve(
        &self,
        point: T::BaseField,
    ) -> Result<T::Affine, HashToCurveError>;
}

Trait for mapping a random field element to a random curve point.

Required Methods

fn new() -> Result<Self, HashToCurveError>

Constructs a new mapping.

fn map_to_curve( &self, point: T::BaseField, ) -> Result<T::Affine, HashToCurveError>

Map an arbitary field element to a corresponding curve point.

Object Safety

This trait is not object safe.