ark_ec::models::mnt6

Type Alias GT

pub type GT<P> = Fp6<P>;

Aliased Type

struct GT<P> {
    pub c0: <Fp6ConfigWrapper<P> as QuadExtConfig>::BaseField,
    pub c1: <Fp6ConfigWrapper<P> as QuadExtConfig>::BaseField,
}

Fields
c0: <Fp6ConfigWrapper<P> as QuadExtConfig>::BaseField

Coefficient c0 in the representation of the field element c = c0 + c1 * X

c1: <Fp6ConfigWrapper<P> as QuadExtConfig>::BaseField

Coefficient c1 in the representation of the field element c = c0 + c1 * X