Struct ark_ec::models::short_weierstrass::Projective

pub struct Projective<P: SWCurveConfig> {
    pub x: P::BaseField,
    pub y: P::BaseField,
    pub z: P::BaseField,
}

Jacobian coordinates for a point on an elliptic curve in short Weierstrass form, over the base field P::BaseField. This struct implements arithmetic via the Jacobian formulae

Fields
x: P::BaseField

X / Z projection of the affine X

y: P::BaseField

Y / Z projection of the affine Y

z: P::BaseField

Projective multiplicative inverse. Will be 0 only at infinity.

Implementations

impl<P: SWCurveConfig> Projective<P>

pub const fn new_unchecked( x: P::BaseField, y: P::BaseField, z: P::BaseField ) -> Self

Constructs a new group element without checking whether the coordinates specify a point in the subgroup.

pub fn new(x: P::BaseField, y: P::BaseField, z: P::BaseField) -> Self

Constructs a new group element in a way while enforcing that points are in the prime-order subgroup.