ark_ec::models::short_weierstrass_jacobian
Struct GroupProjective

pub struct GroupProjective<P: Parameters> {
    pub x: P::BaseField,
    pub y: P::BaseField,
    pub z: P::BaseField,
    /* private fields */
}

Jacobian coordinates for a point on an elliptic curve in short Weierstrass form, over the base field P::BaseField. This struct implements arithmetic via the Jacobian formulae

Fields
x: P::BaseFieldy: P::BaseFieldz: P::BaseField
Implementations

impl<P: Parameters> GroupProjective<P>

pub fn new(x: P::BaseField, y: P::BaseField, z: P::BaseField) -> Self