ark_ec::models

Module twisted_edwards

Structs

Affine

Affine coordinates for a point on a twisted Edwards curve, over the base field P::BaseField.

MontgomeryAffine

Projective

Projective implements Extended Twisted Edwards Coordinates as described in [HKCD08].

Enums

TEFlags

Flags to be encoded into the serialization. The default flags (empty) should not change the binary representation.

Traits

MontCurveConfig

Constants and convenience functions that collectively define the Montgomery model of the curve. In this model, the curve equation is b * y² = x³ + a * x² + x, for constants a and b.

TECurveConfig

Constants and convenience functions that collectively define the Twisted Edwards model of the curve. In this model, the curve equation is a * x² + y² = 1 + d * x² * y², for constants a and d.