ark_ec::models::twisted_edwards

Struct Affine

pub struct Affine<P: TECurveConfig> {
    pub x: P::BaseField,
    pub y: P::BaseField,
}

Affine coordinates for a point on a twisted Edwards curve, over the base field P::BaseField.

Fields
x: P::BaseField

X coordinate of the point represented as a field element

y: P::BaseField

Y coordinate of the point represented as a field element

Implementations

impl<P: TECurveConfig> Affine<P>

pub const fn new_unchecked(x: P::BaseField, y: P::BaseField) -> Self

Construct a new group element without checking whether the coordinates specify a point in the subgroup.

pub fn new(x: P::BaseField, y: P::BaseField) -> Self

Construct a new group element in a way while enforcing that points are in the prime-order subgroup.

pub const fn zero() -> Self

Construct the identity of the group

pub fn is_zero(&self) -> bool

Is this point the identity?

pub fn get_point_from_y_unchecked( y: P::BaseField, greatest: bool, ) -> Option<Self>

Attempts to construct an affine point given an y-coordinate. The point is not guaranteed to be in the prime order subgroup.

If and only if greatest is set will the lexicographically largest x-coordinate be selected.

a * X^2 + Y^2 = 1 + d * X^2 * Y^2 a * X^2 - d * X^2 * Y^2 = 1 - Y^2 X^2 * (a - d * Y^2) = 1 - Y^2 X^2 = (1 - Y^2) / (a - d * Y^2)

pub fn get_xs_from_y_unchecked( y: P::BaseField, ) -> Option<(P::BaseField, P::BaseField)>

Attempts to recover the x-coordinate given an y-coordinate. The resulting point is not guaranteed to be in the prime order subgroup.

If and only if greatest is set will the lexicographically largest x-coordinate be selected.

a * X^2 + Y^2 = 1 + d * X^2 * Y^2 a * X^2 - d * X^2 * Y^2 = 1 - Y^2 X^2 * (a - d * Y^2) = 1 - Y^2 X^2 = (1 - Y^2) / (a - d * Y^2)

pub fn is_on_curve(&self) -> bool

Checks that the current point is on the elliptic curve.

impl<P: TECurveConfig> Affine<P>

pub fn is_in_correct_subgroup_assuming_on_curve(&self) -> bool

Checks if self is in the subgroup having order equaling that of P::ScalarField given it is on the curve.