ark_ec::models::twisted_edwards

Struct Projective

pub struct Projective<P: TECurveConfig> {
    pub x: P::BaseField,
    pub y: P::BaseField,
    pub t: P::BaseField,
    pub z: P::BaseField,
}

Projective implements Extended Twisted Edwards Coordinates as described in [HKCD08].

This implementation uses the unified addition formulae from that paper (see Section 3.1).

Fields
x: P::BaseFieldy: P::BaseFieldt: P::BaseFieldz: P::BaseField
Implementations

impl<P: TECurveConfig> Projective<P>

pub const fn new_unchecked( x: P::BaseField, y: P::BaseField, t: P::BaseField, z: P::BaseField, ) -> Self

Construct a new group element without checking whether the coordinates specify a point in the subgroup.

pub fn new( x: P::BaseField, y: P::BaseField, t: P::BaseField, z: P::BaseField, ) -> Self

Construct a new group element in a way while enforcing that points are in the prime-order subgroup.