ark_ec::models::twisted_edwards

Trait MontCurveConfig

pub trait MontCurveConfig: CurveConfig {
    type TECurveConfig: TECurveConfig<BaseField = Self::BaseField>;

    const COEFF_A: Self::BaseField;
    const COEFF_B: Self::BaseField;
}

Constants and convenience functions that collectively define the Montgomery model of the curve. In this model, the curve equation is b * y² = x³ + a * x² + x, for constants a and b.

Required Associated Constants

const COEFF_A: Self::BaseField

Coefficient a of the curve equation.

const COEFF_B: Self::BaseField

Coefficient b of the curve equation.

Required Associated Types

type TECurveConfig: TECurveConfig<BaseField = Self::BaseField>

Model parameters for the Twisted Edwards curve that is birationally equivalent to this curve.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.