Struct ark_ec::models::twisted_edwards_extended::GroupProjective

pub struct GroupProjective<P: Parameters> {
    pub x: P::BaseField,
    pub y: P::BaseField,
    pub t: P::BaseField,
    pub z: P::BaseField,
    // some fields omitted
}

GroupProjective implements Extended Twisted Edwards Coordinates as described in [HKCD08].

This implementation uses the unified addition formulae from that paper (see Section 3.1).

Fields
x: P::BaseFieldy: P::BaseFieldt: P::BaseFieldz: P::BaseField
Implementations

impl<P: Parameters> GroupProjective<P>

pub fn new(
x: P::BaseField,
y: P::BaseField,
t: P::BaseField,
z: P::BaseField
) -> Self