ark_ec

Module pairing

Structs

MillerLoopOutput

Represents the output of the Miller loop of the pairing.

PairingOutput

Represents the target group of a pairing. This struct is a wrapper around the field that the target group is embedded in.

Traits

Pairing

Collection of types (mainly fields and curves) that together describe how to compute a pairing over a pairing-friendly curve.

Functions

prepare_g1

Preprocesses a G1 element for use in a pairing.

prepare_g2

Preprocesses a G2 element for use in a pairing.