ark_ec
Trait PairingEngine

pub trait PairingEngine:
    Sized
    + 'static
    + Copy
    + Debug
    + Sync
    + Send
    + Eq
    + PartialEq {
    type Fr: PrimeField + SquareRootField;
    type G1Projective: ProjectiveCurve<BaseField = Self::Fq, ScalarField = Self::Fr, Affine = Self::G1Affine> + From<Self::G1Affine> + Into<Self::G1Affine> + MulAssign<Self::Fr>;
    type G1Affine: AffineCurve<BaseField = Self::Fq, ScalarField = Self::Fr, Projective = Self::G1Projective> + From<Self::G1Projective> + Into<Self::G1Projective> + Into<Self::G1Prepared>;
    type G1Prepared: ToBytes + Default + Clone + Send + Sync + Debug + From<Self::G1Affine>;
    type G2Projective: ProjectiveCurve<BaseField = Self::Fqe, ScalarField = Self::Fr, Affine = Self::G2Affine> + From<Self::G2Affine> + Into<Self::G2Affine> + MulAssign<Self::Fr>;
    type G2Affine: AffineCurve<BaseField = Self::Fqe, ScalarField = Self::Fr, Projective = Self::G2Projective> + From<Self::G2Projective> + Into<Self::G2Projective> + Into<Self::G2Prepared>;
    type G2Prepared: ToBytes + Default + Clone + Send + Sync + Debug + From<Self::G2Affine>;
    type Fq: PrimeField + SquareRootField;
    type Fqe: SquareRootField;
    type Fqk: Field;

    // Required methods
    fn miller_loop<'a, I>(i: I) -> Self::Fqk
       where I: IntoIterator<Item = &'a (Self::G1Prepared, Self::G2Prepared)>;
    fn final_exponentiation(_: &Self::Fqk) -> Option<Self::Fqk>;

    // Provided methods
    fn product_of_pairings<'a, I>(i: I) -> Self::Fqk
       where I: IntoIterator<Item = &'a (Self::G1Prepared, Self::G2Prepared)> { ... }
    fn pairing<G1, G2>(p: G1, q: G2) -> Self::Fqk
       where G1: Into<Self::G1Affine>,
             G2: Into<Self::G2Affine> { ... }
}

Required Associated Types

type Fr: PrimeField + SquareRootField

This is the scalar field of the G1/G2 groups.

type G1Projective: ProjectiveCurve<BaseField = Self::Fq, ScalarField = Self::Fr, Affine = Self::G1Affine> + From<Self::G1Affine> + Into<Self::G1Affine> + MulAssign<Self::Fr>

The projective representation of an element in G1.

type G1Affine: AffineCurve<BaseField = Self::Fq, ScalarField = Self::Fr, Projective = Self::G1Projective> + From<Self::G1Projective> + Into<Self::G1Projective> + Into<Self::G1Prepared>

The affine representation of an element in G1.

type G1Prepared: ToBytes + Default + Clone + Send + Sync + Debug + From<Self::G1Affine>

A G1 element that has been preprocessed for use in a pairing.

type G2Projective: ProjectiveCurve<BaseField = Self::Fqe, ScalarField = Self::Fr, Affine = Self::G2Affine> + From<Self::G2Affine> + Into<Self::G2Affine> + MulAssign<Self::Fr>

The projective representation of an element in G2.

type G2Affine: AffineCurve<BaseField = Self::Fqe, ScalarField = Self::Fr, Projective = Self::G2Projective> + From<Self::G2Projective> + Into<Self::G2Projective> + Into<Self::G2Prepared>

The affine representation of an element in G2.

type G2Prepared: ToBytes + Default + Clone + Send + Sync + Debug + From<Self::G2Affine>

A G2 element that has been preprocessed for use in a pairing.

type Fq: PrimeField + SquareRootField

The base field that hosts G1.

type Fqe: SquareRootField

The extension field that hosts G2.

type Fqk: Field

The extension field that hosts the target group of the pairing.

Required Methods

fn miller_loop<'a, I>(i: I) -> Self::Fqk
where I: IntoIterator<Item = &'a (Self::G1Prepared, Self::G2Prepared)>,

Compute the product of miller loops for some number of (G1, G2) pairs.

fn final_exponentiation(_: &Self::Fqk) -> Option<Self::Fqk>

Perform final exponentiation of the result of a miller loop.

Provided Methods

fn product_of_pairings<'a, I>(i: I) -> Self::Fqk
where I: IntoIterator<Item = &'a (Self::G1Prepared, Self::G2Prepared)>,

Computes a product of pairings.

fn pairing<G1, G2>(p: G1, q: G2) -> Self::Fqk
where G1: Into<Self::G1Affine>, G2: Into<Self::G2Affine>,

Performs multiple pairing operations

Object Safety

This trait is not object safe.