Trait ark_ec::hashing::curve_maps::swu::SWUConfig

pub trait SWUConfig: SWCurveConfig {
    const ZETA: Self::BaseField;
}

Trait defining the necessary parameters for the SWU hash-to-curve method for the curves of Weierstrass form of: y^2 = x^3 + a*x + b where ab != 0. From [[WB2019]]

[[WB2019]] https://eprint.iacr.org/2019/403

Required Associated Constants

const ZETA: Self::BaseField

An element of the base field that is not a square root see [WB2019, Section 4]. It is also convenient to have $g(b/ZETA * a)$ to be square. In general we use a ZETA with low absolute value coefficients when they are represented as integers.