Struct ark_ec::models::short_weierstrass_jacobian::GroupAffine

pub struct GroupAffine<P: Parameters> {
    pub x: P::BaseField,
    pub y: P::BaseField,
    pub infinity: bool,
    /* private fields */
}

Affine coordinates for a point on an elliptic curve in short Weierstrass form, over the base field P::BaseField.

Fields
x: P::BaseFieldy: P::BaseFieldinfinity: bool
Implementations

impl<P: Parameters> GroupAffine<P>

pub fn new(x: P::BaseField, y: P::BaseField, infinity: bool) -> Self

pub fn scale_by_cofactor(&self) -> GroupProjective<P>

Multiply self by the cofactor of the curve, P::COFACTOR.

pub fn get_point_from_x(x: P::BaseField, greatest: bool) -> Option<Self>

Attempts to construct an affine point given an x-coordinate. The point is not guaranteed to be in the prime order subgroup.

If and only if greatest is set will the lexicographically largest y-coordinate be selected.

pub fn is_on_curve(&self) -> bool

Checks if self is a valid point on the curve.

pub fn is_in_correct_subgroup_assuming_on_curve(&self) -> bool

Checks if self is in the subgroup having order that equaling that of P::ScalarField.