Struct ark_ec::models::twisted_edwards_extended::GroupAffine

pub struct GroupAffine<P: Parameters> {
    pub x: P::BaseField,
    pub y: P::BaseField,
    /* private fields */
}

Fields
x: P::BaseFieldy: P::BaseField
Implementations

impl<P: Parameters> GroupAffine<P>

pub fn new(x: P::BaseField, y: P::BaseField) -> Self

pub fn scale_by_cofactor(&self) -> <Self as AffineCurve>::Projective

pub fn get_point_from_x(x: P::BaseField, greatest: bool) -> Option<Self>

Attempts to construct an affine point given an x-coordinate. The point is not guaranteed to be in the prime order subgroup.

If and only if greatest is set will the lexicographically largest y-coordinate be selected.

pub fn is_on_curve(&self) -> bool

Checks that the current point is on the elliptic curve.

pub fn is_in_correct_subgroup_assuming_on_curve(&self) -> bool

Checks that the current point is in the prime order subgroup given the point on the curve.