Trait ark_ec::AffineCurve

pub trait AffineCurve: Eq + 'static + Sized + ToBytes + FromBytes + CanonicalSerialize + CanonicalDeserialize + Copy + Clone + Default + Send + Sync + Hash + Debug + Display + Zero + Neg<Output = Self> + Zeroize + Sum<Self> + for<'a> Sum<&'a Self> + From<Self::Projective> {
    type ScalarField: PrimeField + SquareRootField + Into<<Self::ScalarField as PrimeField>::BigInt>;
    type BaseField: Field;
    type Projective: ProjectiveCurve<Affine = Self, ScalarField = Self::ScalarField, BaseField = Self::BaseField> + From<Self> + Into<Self> + MulAssign<Self::ScalarField>;

    const COFACTOR: &'static [u64];

    fn prime_subgroup_generator() -> Self;
    fn from_random_bytes(bytes: &[u8]) -> Option<Self>;
    fn mul<S: Into<<Self::ScalarField as PrimeField>::BigInt>>(
        &self, 
        other: S
    ) -> Self::Projective;
    fn mul_by_cofactor_to_projective(&self) -> Self::Projective;
    fn mul_by_cofactor_inv(&self) -> Self;

    fn into_projective(&self) -> Self::Projective { ... }
    fn mul_by_cofactor(&self) -> Self { ... }
}

Affine representation of an elliptic curve point guaranteed to be in the correct prime order subgroup.

Required Associated Types

type ScalarField: PrimeField + SquareRootField + Into<<Self::ScalarField as PrimeField>::BigInt>

type BaseField: Field

type Projective: ProjectiveCurve<Affine = Self, ScalarField = Self::ScalarField, BaseField = Self::BaseField> + From<Self> + Into<Self> + MulAssign<Self::ScalarField>

Required Associated Constants

const COFACTOR: &'static [u64]

Required Methods

fn prime_subgroup_generator() -> Self

Returns a fixed generator of unknown exponent.

fn from_random_bytes(bytes: &[u8]) -> Option<Self>

Returns a group element if the set of bytes forms a valid group element, otherwise returns None. This function is primarily intended for sampling random group elements from a hash-function or RNG output.

fn mul<S: Into<<Self::ScalarField as PrimeField>::BigInt>>(
&self,
other: S
) -> Self::Projective

Performs scalar multiplication of this element with mixed addition.

fn mul_by_cofactor_to_projective(&self) -> Self::Projective

Multiply this element by the cofactor and output the resulting projective element.

fn mul_by_cofactor_inv(&self) -> Self

Multiply this element by the inverse of the cofactor in Self::ScalarField.

Provided Methods

fn into_projective(&self) -> Self::Projective

Converts self into the projective representation.

fn mul_by_cofactor(&self) -> Self

Multiply this element by the cofactor.